Added education, filter components, pdf generator, recruiter notification for expired jobs

+++ src/app/components/user/projects-carousel/projects-carousel.component.ts
@@ -1,11 +1,13 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { CommonModule, NgFor } from '@angular/common';
-import { MatCardModule } from '@angular/material/card';
-import { MatIcon } from '@angular/material/icon';
-import { ExperienceDto, UserDto } from '../../../services/models';
-import { MatDialog, MatDialogModule } from '@angular/material/dialog';
-import { ExperienceService } from '../../../services/services';
-import { ProjectsFormDialogComponent } from '../projects-form-dialog/projects-form-dialog.component';
+import {Component, Input} from '@angular/core';
+import {CommonModule, NgFor} from '@angular/common';
+import {MatCardModule} from '@angular/material/card';
+import {MatIcon} from '@angular/material/icon';
+import {ExperienceDto, UserDto} from '../../../services/models';
+import {MatDialog, MatDialogModule} from '@angular/material/dialog';
+import {ExperienceService} from '../../../services/services';
+import {ProjectsFormDialogComponent} from '../projects-form-dialog/projects-form-dialog.component';
+import {MatChipsModule} from '@angular/material/chips';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-projects-carousel',
@@ -15,7 +17,8 @@ import { ProjectsFormDialogComponent } from '../projects-form-dialog/projects-fo
     CommonModule,
     NgFor,
     MatIcon,
-    MatDialogModule
+    MatDialogModule,
+    MatChipsModule
   ],
   templateUrl: './projects-carousel.component.html',
   styleUrl: './projects-carousel.component.scss'
@@ -25,8 +28,25 @@ export class ProjectsCarouselComponent {
   @Input() experiencesArray: ExperienceDto[] | undefined;
   currentIndex = 0;
 
+  types = ['All', 'PROJECT', 'COMPETITION'];
+  selectedType = 'All';
+
+  selectType(type: string) {
+    this.selectedType = type;
+    this.currentIndex = 0;
+  }
+
   get experiences(): ExperienceDto[] {
-    return this.experiencesArray?.filter((experience) => experience.type != 'ACCREDITATION') ?? [];
+    switch (this.selectedType) {
+      case 'All':
+        return this.experiencesArray?.filter((experience) => experience.type != 'ACCREDITATION').sort((a, b) => this.parseDate(b.date).getTime() - this.parseDate(a.date).getTime()) ?? [];
+      case 'PROJECT':
+        return this.experiencesArray?.filter((experience) => experience.type === 'PROJECT').sort((a, b) => this.parseDate(b.date).getTime() - this.parseDate(a.date).getTime()) ?? [];
+      case 'COMPETITION':
+        return this.experiencesArray?.filter((experience) => experience.type === 'COMPETITION').sort((a, b) => this.parseDate(b.date).getTime() - this.parseDate(a.date).getTime()) ?? [];
+      default:
+        return this.experiencesArray?.filter((experience) => experience.type != 'ACCREDITATION').sort((a, b) => this.parseDate(b.date).getTime() - this.parseDate(a.date).getTime()) ?? [];
+    }
   };
 
   get emptyExperience(): ExperienceDto {
@@ -41,7 +61,8 @@ export class ProjectsCarouselComponent {
     } as ExperienceDto;
   };
 
-  constructor(public dialog: MatDialog, private experienceService: ExperienceService) { }
+  constructor(public dialog: MatDialog, private experienceService: ExperienceService) {
+  }
 
 
   showNext(): void {
@@ -61,6 +82,7 @@ export class ProjectsCarouselComponent {
       this.currentIndex = this.experiences.length - 1;
     }
   }
+
   showExperience(index: number) {
     this.currentIndex = index;
   }
@@ -68,19 +90,24 @@ export class ProjectsCarouselComponent {
   openDialog(experience: ExperienceDto): void {
     const dialogRef = this.dialog.open(ProjectsFormDialogComponent, {
       width: '50%',
-      data: { user: this.userExperiences, newExperience: experience }
+      data: {user: this.userExperiences, newExperience: experience}
     });
 
     dialogRef.afterClosed().subscribe(result => {
       console.log('Result', result);
       if (result.experience && this.userExperiences) {
+        Swal.fire({
+          icon: "success",
+          title: "Information was successfully saved",
+          showConfirmButton: false,
+          timer: 1500
+        });
         const index = this.experiencesArray?.findIndex(x => x.id === result.experience.id);
         if (index !== -1 && index !== undefined && index !== null && this.experiencesArray !== undefined) {
           this.experiencesArray[index] = result.experience;
           experience = result.experience;
           this.showExperience(index);
-        }
-        else {
+        } else {
           this.experiencesArray?.push(result.experience);
         }
         this.experiencesArray = [...this.experiencesArray ?? []];
@@ -91,33 +118,49 @@ export class ProjectsCarouselComponent {
       console.log('Dialog was closed');
     });
   }
-  deleteExperience(experience: ExperienceDto): void {
-    console.log('Deleting Job History', experience.id);
 
-    if (experience.id !== undefined) {
-      const params = { id: experience.id };
-      this.experienceService.deleteExperience(params).subscribe({
-        next: () => {
-          console.log('Experience deleted successfully');
-          let index = this.experiencesArray?.indexOf(experience);
-          if (index !== undefined && index !== -1) {
-            this.experiencesArray?.splice(index, 1);
-            this.showPrev();
-          }
-        },
-        error: (error) => {
-          console.error('Error deleting experience', error);
-        },
-      });
-    }
+  deleteExperience(experience: ExperienceDto): void {
+    console.log('Deleting Experience', experience.id);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3f51b5",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        if (experience.id !== undefined) {
+          const params = {id: experience.id};
+          this.experienceService.deleteExperience(params).subscribe({
+            next: () => {
+              console.log('Experience deleted successfully');
+              let index = this.experiencesArray?.indexOf(experience);
+              if (index !== undefined && index !== -1) {
+                this.experiencesArray?.splice(index, 1);
+                this.showPrev();
+              }
+            },
+            error: (error) => {
+              console.error('Error deleting experience', error);
+            },
+          });
+        }
+      }
+    });
   }
 
   onImageError(event: any, type: string) {
     if (type === 'PROJECT') {
       event.target.src = 'assets/project.jpg';
-    }
-    else {
+    } else {
       event.target.src = 'assets/competition.jpg';
     }
   }
+
+  private parseDate(dateStr: string): Date {
+    const [day, month, year] = dateStr.split('.').map(Number);
+    return new Date(year, month - 1, day);
+  };
 }
